generate lid on mesh keep immersed later
